Refactoring, code cleaning, separating lex yacc and the main program
